transacciones

+++ src/main/java/com/example/models/Transaccion.java
@@ -28,11 +28,11 @@ public class Transaccion {
 
     // Getters y setters
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
